Сделано онлайн/оффлайн работа с корзиной. Сделана синхронизация локальной корзины с сервером. TODO сделать синхронизацию локальной корзины с сервером через WorkManager (android) и хз-что (ios)

+++ composeApp/src/commonMain/kotlin/com/foodsaver/app/common/ProductCounter.kt
@@ -29,7 +29,7 @@ import org.jetbrains.compose.resources.vectorResource
 
 @Composable
 fun ProductCounter(
-    productCount: Int,
+    productCount: Long,
     onIncreaseClick: () -> Unit,
     onDecreaseClick: () -> Unit
 ) {
